fix(security): close audit findings — tar/zip-slip, shell injection, panics, perms
Multi-dimension hardening sweep (security, robustness, concurrency, supply-chain).

Security:
- archive: containment checks fix critical tar-slip + zip-slip (CWE-22),
  symlink/hardlink escape (CWE-59); add decompression-bomb caps
- forloop/task: close shell command injection on POSIX (env/argv) and Windows
  (cmd /V:ON delayed expansion !var!); pinned by regression tests
- crypto: PBKDF2 minimum-iteration floor; sqlite identifier quoting; 0600 perms
  on profile/config/key files; curl -k now actually wired; redirect/SSRF +
  timeout guards; goja interrupt; bounded remote io.ReadAll
Robustness:
- fix panics in ls/ps/dd/free/yq; top-level panic recovery with deterministic
  exit code; atomic profile writes; bbolt alloc bound; channel pagination bound
Concurrency:
- rg context-cancel deadlock + goroutine leak; twig progress-callback race;
  pipeline errors.Is; gopsagent connection cap + read deadlines
Supply chain:
- -trimpath in .goreleaser.yaml + Taskfile.yml for reproducible builds
- uname_windows reimplemented pure-Go (RtlGetVersion, no exec)

Deferred/accepted (see docs/quality/HARDENING.md "Resolution status"):
- crypto-02: PBKDF2 default bump needs a versioned ciphertext envelope first
- machineid_darwin: keeps ioreg (no pure-Go path to same UUID without bricking
  existing master.key) — documented machine-identity exec exception
- 8 sanctioned exec-wrapper commands documented in docs/architecture/patterns.md

Verified: go build/vet, GOOS=darwin/windows cross-compile, gofmt, golangci-lint,
and targeted tests all green.

+++ internal/cli/ls/ls_unix.go
@@ -9,9 +9,12 @@ import (
 
 func fillUnixInfo(entry *Entry, info fs.FileInfo) {
 	if stat, ok := info.Sys().(*syscall.Stat_t); ok {
-		entry.Inode = stat.Ino
-		entry.NLink = stat.Nlink
-		entry.UID = stat.Uid
-		entry.GID = stat.Gid
+		// syscall.Stat_t field widths vary by GOOS/GOARCH (e.g. Nlink is
+		// uint16 on darwin, uint64 on linux/amd64, int* on some BSDs).
+		// Convert explicitly so this compiles on every supported platform.
+		entry.Inode = uint64(stat.Ino)
+		entry.NLink = uint64(stat.Nlink)
+		entry.UID = uint32(stat.Uid)
+		entry.GID = uint32(stat.Gid)
 	}
 }
